pyproject.toml
[tool.poetry]
name = "course-work-5"
version = "0.1.0"
description = ""
authors = ["Aleksandr Charushin <charushinas@gmail.com>"]
readme = "README.md"
packages = [{include = "course_work_5"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
